Fixed some cast warnings
Also caught a typo of the default label in console.cpp .

Updated assets.

+++ include/physics-system.hpp
@@ -96,7 +96,8 @@ namespace tec {
 		*/
 		void Update(const double delta);
 
-		eid RayCastMousePick(eid source_entity, float mouse_x = 0.0f, float mouse_y = 0.0f, float screen_width = 1.0f, float screen_height = 1.0f);
+		eid RayCastMousePick(eid source_entity, double mouse_x = 0.0f, double mouse_y = 0.0f,
+			float screen_width = 1.0f, float screen_height = 1.0f);
 		eid RayCastIgnore(eid);
 		glm::vec3 GetLastRayPos() const {
 			btVector3 tmp = last_raypos; // grab a copy
